Extract Habits.get_longest_streak
