+++ lib/habits/achievements/streak_length.ex
@@ -4,8 +4,9 @@ defmodule Habits.Achievements.StreakLength do
   use Habits.Achievements.Achievement
 
   alias Habits.Habits.Habit
+  alias Habits.Habits
 
   defp value_for(%Habit{} = habit) do
-    Habit.get_longest_streak(habit)
+    Habits.get_longest_streak(habit)
   end
 end
